+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/markers/config/__init__.py
@@ -1,12 +1,52 @@
+# Copyright (c) 2022, NVIDIA CORPORATION & AFFILIATES, ETH Zurich, and University of Toronto
+# All rights reserved.
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
 from __future__ import annotations
 
+import omni.isaac.orbit.sim as sim_utils
+from omni.isaac.orbit.markers.visualization_markers import VisualizationMarkersCfg
 from omni.isaac.orbit.utils.assets import ISAAC_NUCLEUS_DIR
 
-from ..visualization_markers import VisualizationMarkersCfg
+##
+# Sensors.
+##
+
+RAY_CASTER_MARKER_CFG = VisualizationMarkersCfg(
+    markers={
+        "hit": sim_utils.SphereCfg(
+            radius=0.02,
+            visual_material=sim_utils.PreviewSurfaceCfg(diffuse_color=(1.0, 0.0, 0.0)),
+        ),
+    }
+)
+"""Configuration for the ray-caster marker."""
+
+
+CONTACT_SENSOR_MARKER_CFG = VisualizationMarkersCfg(
+    markers={
+        "contact": sim_utils.SphereCfg(
+            radius=0.02,
+            visual_material=sim_utils.PreviewSurfaceCfg(diffuse_color=(1.0, 0.0, 0.0)),
+        ),
+        "no_contact": sim_utils.SphereCfg(
+            radius=0.02,
+            visual_material=sim_utils.PreviewSurfaceCfg(diffuse_color=(0.0, 1.0, 0.0)),
+            visible=False,
+        ),
+    },
+)
+"""Configuration for the contact sensor marker."""
+
+
+##
+# Frames.
+##
 
 FRAME_MARKER_CFG = VisualizationMarkersCfg(
     markers={
-        "frame": VisualizationMarkersCfg.FileMarkerCfg(
+        "frame": sim_utils.UsdFileCfg(
             usd_path=f"{ISAAC_NUCLEUS_DIR}/Props/UIElements/frame_prim.usd",
             scale=(0.5, 0.5, 0.5),
         )
@@ -14,73 +54,71 @@ FRAME_MARKER_CFG = VisualizationMarkersCfg(
 )
 """Configuration for the frame marker."""
 
+
+RED_ARROW_X_MARKER_CFG = VisualizationMarkersCfg(
+    markers={
+        "arrow": sim_utils.UsdFileCfg(
+            usd_path=f"{ISAAC_NUCLEUS_DIR}/Props/UIElements/arrow_x.usd",
+            scale=(1.0, 0.1, 0.1),
+            visual_material=sim_utils.PreviewSurfaceCfg(diffuse_color=(1.0, 0.0, 0.0)),
+        )
+    }
+)
+"""Configuration for the red arrow marker (along x-direction)."""
+
+
+BLUE_ARROW_X_MARKER_CFG = VisualizationMarkersCfg(
+    markers={
+        "arrow": sim_utils.UsdFileCfg(
+            usd_path=f"{ISAAC_NUCLEUS_DIR}/Props/UIElements/arrow_x.usd",
+            scale=(1.0, 0.1, 0.1),
+            visual_material=sim_utils.PreviewSurfaceCfg(diffuse_color=(0.0, 0.0, 1.0)),
+        )
+    }
+)
+"""Configuration for the blue arrow marker (along x-direction)."""
+
+GREEN_ARROW_X_MARKER_CFG = VisualizationMarkersCfg(
+    markers={
+        "arrow": sim_utils.UsdFileCfg(
+            usd_path=f"{ISAAC_NUCLEUS_DIR}/Props/UIElements/arrow_x.usd",
+            scale=(1.0, 0.1, 0.1),
+            visual_material=sim_utils.PreviewSurfaceCfg(diffuse_color=(0.0, 1.0, 0.0)),
+        )
+    }
+)
+"""Configuration for the green arrow marker (along x-direction)."""
+
+
+##
+# Goals.
+##
+
+CUBOID_MARKER_CFG = VisualizationMarkersCfg(
+    markers={
+        "cuboid": sim_utils.CuboidCfg(
+            size=(0.1, 0.1, 0.1),
+            visual_material=sim_utils.PreviewSurfaceCfg(diffuse_color=(1.0, 0.0, 0.0)),
+        ),
+    }
+)
+"""Configuration for the cuboid marker."""
+
 POSITION_GOAL_MARKER_CFG = VisualizationMarkersCfg(
     markers={
-        "target_far": VisualizationMarkersCfg.MarkerCfg(
-            prim_type="Sphere",
-            color=(1.0, 0.0, 0.0),
-            attributes={"radius": 0.01},
+        "target_far": sim_utils.SphereCfg(
+            radius=0.01,
+            visual_material=sim_utils.PreviewSurfaceCfg(diffuse_color=(1.0, 0.0, 0.0)),
         ),
-        "target_near": VisualizationMarkersCfg.MarkerCfg(
-            prim_type="Sphere",
-            color=(0.0, 1.0, 0.0),
-            attributes={"radius": 0.01},
+        "target_near": sim_utils.SphereCfg(
+            radius=0.01,
+            visual_material=sim_utils.PreviewSurfaceCfg(diffuse_color=(0.0, 1.0, 0.0)),
         ),
-        "target_invisible": VisualizationMarkersCfg.MarkerCfg(
-            prim_type="Sphere",
-            color=(0.0, 0.0, 1.0),
-            attributes={"radius": 0.01},
+        "target_invisible": sim_utils.SphereCfg(
+            radius=0.01,
+            visual_material=sim_utils.PreviewSurfaceCfg(diffuse_color=(0.0, 0.0, 1.0)),
             visible=False,
         ),
     }
 )
 """Configuration for the end-effector tracking marker."""
-
-RAY_CASTER_MARKER_CFG = VisualizationMarkersCfg(
-    markers={
-        "hit": VisualizationMarkersCfg.MarkerCfg(
-            prim_type="Sphere",
-            color=(1.0, 0.0, 0.0),
-            attributes={"radius": 0.02},
-        ),
-    },
-)
-"""Configuration for the ray-caster marker."""
-
-
-CONTACT_SENSOR_MARKER_CFG = VisualizationMarkersCfg(
-    markers={
-        "contact": VisualizationMarkersCfg.MarkerCfg(
-            prim_type="Sphere",
-            color=(1.0, 0.0, 0.0),
-            attributes={"radius": 0.02},
-        ),
-        "no_contact": VisualizationMarkersCfg.MarkerCfg(
-            visible=False,
-            prim_type="Sphere",
-            color=(0.0, 1.0, 0.0),
-            attributes={"radius": 0.02},
-        ),
-    },
-)
-"""Configuration for the contact sensor marker."""
-
-
-ARROW_X_MARKER_CFG = VisualizationMarkersCfg(
-    markers={
-        "arrow": VisualizationMarkersCfg.FileMarkerCfg(
-            usd_path=f"{ISAAC_NUCLEUS_DIR}/Props/UIElements/arrow_x.usd",
-            scale=[1.0, 0.1, 0.1],
-        )
-    }
-)
-"""Configuration for the arrow marker (along x-direction)."""
-
-CUBOID_MARKER_CFG = VisualizationMarkersCfg(
-    markers={
-        "cuboid": VisualizationMarkersCfg.MarkerCfg(
-            prim_type="Cube",
-        )
-    }
-)
-"""Configuration for the cuboid marker."""