+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/sim/spawners/from_files/from_files_cfg.py
@@ -8,9 +8,9 @@ from __future__ import annotations
 from dataclasses import MISSING
 from typing import Callable
 
-from omni.isaac.orbit.sim import loaders
+from omni.isaac.orbit.sim import loaders, schemas
 from omni.isaac.orbit.sim.spawners import materials
-from omni.isaac.orbit.sim.spawners.spawner_cfg import ArticulationSpawnerCfg, SpawnerCfg
+from omni.isaac.orbit.sim.spawners.spawner_cfg import RigidObjectSpawnerCfg, SpawnerCfg
 from omni.isaac.orbit.utils import configclass
 from omni.isaac.orbit.utils.assets import ISAAC_NUCLEUS_DIR
 
@@ -18,7 +18,37 @@ from . import from_files
 
 
 @configclass
-class UsdFileCfg(ArticulationSpawnerCfg):
+class FileCfg(RigidObjectSpawnerCfg):
+    """Configuration parameters for spawning an asset from a file.
+
+    Note:
+        By default, all properties are set to None. This means that no properties will be added or modified
+        to the prim outside of the properties available by default when spawning the prim.
+    """
+
+    scale: tuple[float, float, float] | None = None
+    """Scale of the asset. Defaults to None, in which case the scale is not modified."""
+
+    articulation_props: schemas.ArticulationPropertiesCfg | None = None
+    """Properties to apply to the articulation root."""
+
+    visual_material_path: str = "material"
+    """Path to the visual material to use for the prim. Defaults to "material".
+
+    If the path is relative, then it will be relative to the prim's path.
+    This parameter is ignored if `visual_material` is not None.
+    """
+
+    visual_material: materials.VisualMaterialCfg | None = None
+    """Visual material properties to override the visual material properties in the URDF file.
+
+    Note:
+        If None, then no visual material will be added.
+    """
+
+
+@configclass
+class UsdFileCfg(FileCfg):
     """USD file to spawn asset from.
 
     See :meth:`spawn_from_usd` for more information.
@@ -35,7 +65,7 @@ class UsdFileCfg(ArticulationSpawnerCfg):
 
 
 @configclass
-class UrdfFileCfg(ArticulationSpawnerCfg, loaders.UrdfLoaderCfg):
+class UrdfFileCfg(FileCfg, loaders.UrdfLoaderCfg):
     """URDF file to spawn asset from.
 
     It uses the :class:`UrdfLoader` class to create a USD file from URDF and spawns the imported